Reset LOCAL_PATH so that `mma` can work properly
The value of LOCAL_PATH is overridden by "include $(art_path)/...".
This causes LOCAL_PATH to be empty string when registering modules in
art/Android.mk and this causes MODULES-IN-* logic to not register
properly, hence `mma` not work properly.
Reset the value of LOCAL_PATH back to $(art_path) so that base_rules.mk
can record the module path properly.

Bug: 162918277
Test: m com.android.art
Test: m MODULES-IN-art
Change-Id: I79c5becd0cb81ed85bb5043d5b8150d550657db9