ART: Add IsInputNonNegativeAt and IsInputNonNegativeOrMinIntAt to HInstruction
When it can be quickly checked that an input operand in non-negative, additional
optimizations can be applied during code generation.

The CL adds IsInputNonNegativeAt and IsInputNonNegativeOrMinIntAt to HInstruction
which can be used to check if the input operant at the index is
non-negative. They guarantee that at the time of checks the HInstruction
can have non-negative inputs. Other optimizations after that might break
the invariant.

Optimizations HRem/HDiv for ARM32/ARM64 are moved to used the new methods.

Test: 411-checker-hdiv-hrem-pow2
Test: 411-checker-hdiv-hrem-const
Test: test.py --host --optimizing --jit --gtest --interpreter
Test: test.py -target --optimizing --jit --interpreter
Test: run-gtests.sh
Change-Id: Icf8574699e003bba194097c4e39660de16aa53d9