Make 1932 robust against unexpectedly inflated monitors
Test 1932 checks that the JVMTI monitor events works. As an
implementation detail in ART we consider any attempt to lock an
inflated monitor contended. Since monitors can end up inflated for a
number of reasons and stay inflated for indeterminate periods this
means that we might sometimes get events when we don't expect them.
This could cause test 1932 to flake. To reduce this flakiness use
internal APIs to force monitors into deflated states when possible.

Test: ./test.py --host
Bug: 149308087
Change-Id: I66873453231eedfe69aaebfe6a59e04f7cd77229