Use NotifyDatasourceStopped for perfetto_hprof.
This takes the guesswork out of how long we need to run the trace for.
Now, we can have a short trace duration (but long enough for the fork +
datasource init to happen), and a long datasource_stop_timeout. This
way, the trace will only take as long as it needs to.

Bug: 159594916
Change-Id: I0fa4219f454e91476833297342a11009e0e27435