[metrics] Define Histogram
Adds a definition of the Histogram metric type. This CL also declares
the JitMethodCompileTime histogram, but collection for this metric is
not yet included.

The ArtMetrics class is also refactored so that metrics are not exposed
directly as fields, but instead through accessor functions.

Finally, this CL includes a test case for ArtMetrics::ReportAllMetrics
that demonstrates how the set of metrics are reported to a backend.

Bug: 170149255
Test: m test-art-host-gtest-art_libartbase_tests
Change-Id: I5041588609d9c386ce053c16713c9e9bb5210e3a