Override maxTargetSdk for @UnsupportedAppUsage with special bug id
We are planning to restrict a number of APIs that we don't think
are used. We will temporarily treat them as if they were in the
blocklist in order to give us early warning. They will be moved
to maxTargetSdk=R before the final release.

This bug with special treatment is 170729553

Test: atest UnsupportedAppUsageAnnotationHandlerTest
Fixes: 170943969
Change-Id: I96c196eb33303fa70b6b179285338ccb5d14d78f