Use Soong config variables to disable modules that conflict with ART Module prebuilts.
When ART Module prebuilts are in use (i.e. have prefer:true), they
interfere apex/sdk/module_exports modules and cause Soong failures. To
work around that problem we disable the modules by default and only
enable them when ART-specific products are used which should be used
with the ART sources. By leaving them disabled by default, platform
building will work from ART prebuilts even when the rest of the ART
sources are present.

This topic does all the preparations for that except the actual
disabling, which has to be done at the same time prebuilts are taken
into use.

Test: m SKIP_BOOT_JARS_CHECK=true droid
  on AOSP master with ART SDKs and APEXes disabled by default and
  prebuilt ART release and debug APEXes present
Test: art/build/build-art-module.sh
      art/build/apex/runtests.sh
      art/tools/buildbot-build.sh {--host,--target}
      art/tools/run-gtests.sh (in device chroot)
      art/test/testrunner/testrunner.py --target --64 --optimizing
        (in device chroot)
  on master-art with ART SDKs and APEXes disabled by default and
  verified that they are still enabled
Bug: 172480617
Change-Id: I1126e7dc40ab55fca1df0043199da5cb449de14d