Set apex_available for signal_dumper
The module should be made available to the platform because the platform
variant is required in Android.run-test.mk. Note that installing the
module to the platform is still impossible (which is good) as
LOCAL_UNINSTALLABLE_MODULE is set to true by the build system. This
change just makes Android-<target>.mk to have an entry for this module
so that places like Android.run-test.mk, etc don't fail.

Bug: 174333583
Test: watch TH for master-art-host targets
Change-Id: I8d6009093da8a883595db81cab67c8b3320bc9ca