Fix MessageQueueTests
This test requires a set of fixed messages to come in order, but there
is one TimeoutExpiredMessage that can arrive at any time. Originally,
the logic for this test was incorrect, as it would allow any message to
be replaced by a timeout message, but then the sequence would be out of
sync. The change now explicitly advances the message pointer, which
means the timeout message does not effectively replace an expected
message.

This fixes another issue where MessageQueue may try to call
ConditionVariable::TimedWait with a negative timeout, which fails with
an EINVAL error during the call to futex.

Test: m test-art-host-gtest-art_runtime_tests
Bug: 174652565
Change-Id: Ib4b6d63125e1910b137729cefda0b894cbaa4bf0