[metrics] Add file output support
Adds a new command line option, -Xwrite-metrics-to-file=_, which
specifies a file for writing metrics to. This can be used in conjunction
with -Xwrite-metrics-to-log, which will cause the output to go to both
logcat and the file.

Writing metrics to a file can be helpful for testing or benchmarking
environments.

Access to the file is protected using advisory locking, so it is safe to
have multiple processes write metrics to the same file.

Test: test/run-test --host test/2234-write-metrics-to-file
Bug: 175025360, 170149255
Change-Id: I83a4f959cbb96847dc71188035c050a73b915064