Enable LSE of shadow-klass field
The Object.shadow$_klass_ field is special in that it has a non-null
default value. Previously, however, LSE would always treat it as
having an Unknown value. This meant LSE was unable to handle reads
from the shadow$_klass_ field and replace it with the associated
LoadClass, leading to some allocations not being eliminated.

This change correctly informs LSE that the value should initially be
the LoadClass instruction.

Test: ./test.py --host
Bug: 67037140

Change-Id: I9099001211c6c099a5d1ab4188e2b11f79e2ad9e