Adjust min_free down for zygote
The working heap of a zygote process should be as small as possible.
By garbage collecting more frequently, we risk dirtying the heap.
But since this turns into the child's Java heap, it's expected that
the child will dirty it anyway, and there will be minimal sharing
with the zygote in any case.

This will slow down the early phases of the zygote, before zygote
heap creation, a bit, but that shouldn't be very noticeable,
especially since we shouldn't collect during the first 2 seconds.

Test: Built and booted AOSP.
Bug: 174211442
Change-Id: Icad9ef89ecc79136854bb5bf938642f537d7d505