Fix to not copy pseudo-random module variants to art-test-cases.zip.
The install hook can be called for more than one module variant for the
same library (for host). In that case it would just copy the last one,
which may be some sanitizer variant that won't work.

The install hook needs to check the SkipInstall flag to pick the right
variant.

This patch also limits the copying to the BuildOs variants, since the
OsClass "host" is used for Windows as well.

Also error out if conflicting source paths are detected, so this error
can't slip by unnoticed.

Test: m nothing
Test: m art-host-tests
      unzip out/target/product/vsoc_x86/art-host-tests.zip
      host/testcases/art_common/out/host/linux-x86/bin/dexdiag
  Check that dexdiag runs without linker symbol error.
Test: Check that testcasesContent() returns the same number of entries
  before and after the IsSkipInstall() filter.
Change-Id: Ief8168e223fc196a49fa266aae38fead71e574ae