Use HashMap<> and HashSet<> in ClassLinker.
Use HashMap<> and HashSet<> instead of std::unordered_map<>
and std::unordered_set<> as it is much faster thanks to
avoiding individual nodes that need malloc/free. Though this
can increase memory usage in some cases, these particular
containers are short-lived and not that big.

Move the definition of ClassLinker::MethodTranslation from
class_linker.h to class_linker.cc. Add a default constructor
used for creating empty entries in HashMap<>.

Fix HashMap<> implementation for std::pair<> being used
as the Key. Make Signature copy-assignable so that we can
use it as (part of) a Value in HashMap<>.

Also avoid unnecessary std::vector<> for FillIfTable().

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Test: boots.
Bug: 175869411
Change-Id: I48722fb3bb7780bf12fb71c7449494e097e4a368