Revert^2 "[metrics] Add file output support"
Adds a new command line option, -Xwrite-metrics-to-file=_, which
specifies a file for writing metrics to. This can be used in conjunction
with -Xwrite-metrics-to-log, which will cause the output to go to both
logcat and the file.

Writing metrics to a file can be helpful for testing or benchmarking
environments.

Access to the file is protected using advisory locking, so it is safe to
have multiple processes write metrics to the same file.

This reverts commit ce2e0957f065002293ef152da28b6fa941c6baa3.

Reason for revert: Fixed test failures by adding Android.bp file.

Change-Id: I6768f5b4920896fc0f8e5866c878af999f73012d
Test: test/run-test --host test/2234-write-metrics-to-file
Bug: 175025360
Bug: 170149255