A few file descriptor sanity clean-ups
Use unique_fd in oat_file_assistant. Warnings were emitted from
dexoptanalyzer about assuming a file was okay, the root cause of this
warning was not being able to open the zip file. This was due to the
file descriptor being closed when the jar/apk/zip file was
opened. There was a deeply nested assumption about taking ownership.

Update dexoptanalyzer to use unique_fd and also check passed file
descriptor is open, not just a valid numeral.

Test: m art
Test: manual dexoptanalyzer observe log output
Change-Id: I6311dacbc1f3f452a4fd6e9a36ecd7aa3c5e403b