[metrics] Add histogram statistics for reporting
In order to meet feature parity with the existing GC histograms, this
CL adds the ability to estimate percentiles and confidence intervals
from histogram data.

This is currently exposed only to metrics backends, which can choose
to compute these summary statistics if desired. The human-readable
stream backend does this.

Example output:

    JitMethodCompileTime: range = 0...1000000
      99% confidence interval: 333...66332
      buckets: 1,0,0,0,0,0,0,0,0,0,0,0,0,0,0

Bug: 178034939
Test: m test-art-host-gtest-art_runtime_tests
Change-Id: Ia0f8f53419018583af20029c4c9e1ea9d59a9985