Silence `-Wunguarded-availability` in `libartbase(d)` on macOS.
The build system recently started to use
`-Wunguarded-availability` and is complaining that on macOS the
calls to `clock_gettime` (introduced in macOS 10.12) are not
properly guarded in `libartbase/base/time_utils.cc` with the
`__builtin_availability` runtime check when the deployment target
is macOS 10.10.0. Silence this warning as we don't intend to
support macOS versions that do not provide `clock_gettime`.

Change-Id: I5f4effe35990b2b173d70fe9571e435b304545b0
Test: macOS build with Forrest containing this change and
      https://android-review.googlesource.com/c/platform/external/libusb/+/1556459/
Bug: 178124881