arm: Fix and improve JNI frame entry/exit.
Add workaround for vixl emitting UNPREDICTABLE instruction
in frame exit. Pop LR directly to PC when possible and emit
previously missing CFI data otherwise.

For frames that just save LR and align the stack with one
more unused word, make the frame entry a single PUSH (by
arbitrarily storing r3) and make exit a single POP (by
arbitrarily restoring r3 which we are allowed to clobber
as it's a caller-save register not used for return value).

Test: run-gtests.sh
Test: testrunner.py --target --32 --optimizing
Bug: 178048807
Change-Id: I8752711bfd4bcfc2c1cfa371108de393df00aeb8