Ignore corrupt vdex files from .dm
Perform an earlier validation of the .vdex files when they come
from .dm and if they turn out to be out of date or corrupt, ignore
them. The compilation can proceed without them.

This will accomodate cases like the .vdex being out of date or
build with a previous version of the compiler.

Test: gtest
Bug: 177462384
Change-Id: Ib1ad0f86950877cedbb57668767b6acb884c3227