[metrics] Turn on statsd reporting by default
This change enables statsd reporting by default, which means ART will
always report startup measurements to statsd.

In the future, this will normally be controlled through experiments,
but this will let us get some early insight to how ART's metrics
reporting affects statsd.

This is a single line CL so it should be easy to revert if it causes
problems.

Test: m
Bug: 170149255
Change-Id: I77d1f01d056e1c8e823bf2802ae987e350457911