Add `MainlineTestModuleController` support for AOSP to ART tests.
Since the introduction of `MainlineTestModuleController` support in
ART run-tests and gtests, when these tests were run in a Mainline
context (e.g. with `mts-tradefed`), they were only enabled when the
internal version of the ART Mainline Module (`com.google.android.art`)
was present on the device used for testing. However, AOSP platform and
module builds contain the external version of the ART Mainline
Module (`com.android.art`). Allow ART tests to run in Mainline testing
when either of these modules is installed.

Test: m mts && mts-tradefed run commandAndExit mts-art (with AOSP build)
Bug: 167385698
Change-Id: I543aa6b51c4812bbaba3352c9c241e5ee898b783