Trigger fewer GCs during startup
WIP: DO NOT SUBMIT

Instead of explicitly triggering a GC after two seconds, gradually
reduce the GC triggering threshold. In particular, a small process
that almost immediately goes into the background should GC only
as part of the transition to background.

Clarify the concurrency rules for accessing concurrent_start_bytes_
in the process.

Test: Mostly builds and then crashes in dex2oat.
Bug: 181351667
Change-Id: Id85d7e405b9305a70ba751a82ab7c2226eb83308